decomposition in websocket/; front dockerfile fixed; custom migration with inserts
