+++ frontend/Dockerfile
@@ -2,7 +2,10 @@ FROM node:19-alpine3.16
 
 WORKDIR /frontend
 COPY package.json .
-RUN npm install
+
+RUN npm config set fetch-retry-mintimeout 20000
+RUN npm config set fetch-retry-maxtimeout 120000
+RUN npm install -g npm lerna cross-env rimraf --loglevel notice
 
 COPY . ./
